some small refactor plus changing the start up command
