+++ README.md
@@ -5,4 +5,8 @@ playing around with mcp servers and aks
 # set up
 Need to document the setup
 
+Run with Inspector
 `uv run mcp dev aks.py`
+
+
+You can also use 5ire to connect with local
